// src/generated/api/model/runConfig.ts
/**
 * Registrar Backend API
 * Manages the Alloy Analyzer & the model in the background.
 *
 * The version of the OpenAPI document: 0.1.9
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SemConfig } from './semConfig';


export interface RunConfig { 
    genEdProgram?: string;
    pathway?: Array<SemConfig>;
}